requirements.txt
SQLAlchemy==2.0.36
PyJWT==2.10.1
pydantic==2.10.3
fastapi==0.115.6
uvicorn==0.32.1
websockets==14.1
aiohttp==3.11.10
psycopg2-binary==2.9.10
Jinja2==3.1.4
python-logging-loki==0.3.1